FEATURE: Add limit for trial
Before this, user can input invalid option for unlimited of times.
Lecturer do not like this. He think it is not good, although I do not
know why.

Now a limit of trial is applied. If user enter invalid option for
more than 3 times, the user will redirected back to main menu.

This change only apply to options input. It does not apply to text
input.