REFACTOR: Change interface of `get_choice_from_user`
Before this, `get_choice_from_user` does not support showing disabled
options, this function is needed.

`options` parameter of `get_choice_from_user` is modified, from a list
of string, to list of dict, which contains 'key', 'value', and an
optional 'disabled' keys. If there exists a 'disabled' key and it is
`True`, then that option will still be printed but user cannot select
that option.

Since the definition of `get_choice_from_user` is changed, other part of
code that call it has been modified too.